+++ src/shapes/text.js
@@ -4,6 +4,17 @@ var Styles = require('../mixins/styles');
 var Utils = require('../utils');
 var svg = require('virtual-dom/virtual-hyperscript/svg');
 
+var map = {
+  textAlign: 'text-align',
+  fontFamily: 'font-family',
+  fontStyle: 'font-style',
+  fontWeight: 'font-weight',
+  fontSize: 'font-size',
+  letterSpacing: 'letter-spacing',
+  textDecoration: 'text-decoration'
+};
+var keys = Object.keys(map);
+
 var Text = function(text, x, y) {
   this.shape();
   this.styles();
@@ -18,58 +29,12 @@ Text.prototype = {
     throw new Error('You need the Rune.Font plugin to convert text to polygon');
   },
 
-  textAlign: function(textAlign) {
-    this.state.textAlign = textAlign;
-    this.changed();
-    return this;
-  },
-
-  fontFamily: function(fontFamily) {
-    this.state.fontFamily = fontFamily;
-    this.changed();
-    return this;
-  },
-
-  fontStyle: function(fontStyle) {
-    this.state.fontStyle = fontStyle;
-    this.changed();
-    return this;
-  },
-
-  fontWeight: function(fontWeight) {
-    this.state.fontWeight = fontWeight;
-    this.changed();
-    return this;
-  },
-
-  fontSize: function(fontSize) {
-    this.state.fontSize = fontSize;
-    this.changed();
-    return this;
-  },
-
-  letterSpacing: function(letterSpacing) {
-    this.state.letterSpacing = letterSpacing;
-    this.changed();
-    return this;
-  },
-
-  textDecoration: function(textDecoration) {
-    this.state.textDecoration = textDecoration;
-    this.changed();
-    return this;
-  },
-
   copy: function(parent) {
     var copy = new Text();
     copy.state.text = this.state.text;
-    copy.state.textAlign = this.state.textAlign;
-    copy.state.fontFamily = this.state.fontFamily;
-    copy.state.fontStyle = this.state.fontStyle;
-    copy.state.fontWeight = this.state.fontWeight;
-    copy.state.fontSize = this.state.fontSize;
-    copy.state.letterSpacing = this.state.letterSpacing;
-    copy.state.textDecoration = this.state.textDecoration;
+    for (var i = 0; i < keys.length; i++) {
+      copy.state[keys[i]] = this.state[keys[i]];
+    }
     Utils.copyMixinVars(this, copy);
     Utils.groupLogic(copy, this.parent, parent);
     return copy;
@@ -90,29 +55,26 @@ Text.prototype = {
     this.shapeAttributes(attr);
     this.stylesAttributes(attr);
 
-    // attributes that need specific handling
-    if (this.state.textAlign) {
-      var translate = { left: 'start', center: 'middle', right: 'end' };
-      attr['text-anchor'] = translate[this.state.textAlign];
+    for (var i = 0; i < keys.length; i++) {
+      if (this.state[keys[i]]) {
+        attr[map[keys[i]]] = Utils.s(this.state[keys[i]]);
+      }
     }
 
-    if (this.state.fontFamily)
-      attr['font-family'] = Utils.s(this.state.fontFamily);
-    if (this.state.textAlign)
-      attr['text-align'] = Utils.s(this.state.textAlign);
-    if (this.state.fontStyle)
-      attr['font-style'] = Utils.s(this.state.fontStyle);
-    if (this.state.fontWeight)
-      attr['font-weight'] = Utils.s(this.state.fontWeight);
-    if (this.state.fontSize) attr['font-size'] = Utils.s(this.state.fontSize);
-    if (this.state.letterSpacing)
-      attr['letter-spacing'] = Utils.s(this.state.letterSpacing);
-    if (this.state.textDecoration)
-      attr['text-decoration'] = Utils.s(this.state.textDecoration);
+    // handle textalign conversion
+    if (attr['text-align']) {
+      var translate = { left: 'start', center: 'middle', right: 'end' };
+      attr['text-anchor'] = translate[attr['text-align']];
+    }
 
     return svg('text', attr, this.state.text);
   }
 };
+
+// Generate setters
+for (var i = 0; i < keys.length; i++) {
+  Text.prototype[keys[i]] = Utils.getSetter(keys[i]);
+}
 
 assign(Text.prototype, Shape, Styles, { type: 'text' });